+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/ConfigurationTest.php
@@ -665,6 +665,89 @@ class ConfigurationTest extends TestCase
         $this->assertSame([['place' => 'c', 'weight' => 1]], $transitions[1]['to']);
     }
 
+    public function testWorkflowEventsToDispatchRejectsMixedAllowListAndBlockList()
+    {
+        $processor = new Processor();
+        $configuration = new Configuration(true);
+
+        $this->expectException(InvalidConfigurationException::class);
+        $this->expectExceptionMessage('Cannot mix allow-list and block-list entries in "events_to_dispatch": every entry must start with "!" (block-list mode) or none of them must (allow-list mode).');
+
+        $processor->processConfiguration($configuration, [[
+            'http_method_override' => false,
+            'handle_all_throwables' => true,
+            'php_errors' => ['log' => true],
+            'workflows' => [
+                'workflows' => [
+                    'mixed' => [
+                        'supports' => [self::class],
+                        'places' => ['a', 'b'],
+                        'initial_marking' => 'a',
+                        'events_to_dispatch' => ['workflow.enter', '!workflow.announce'],
+                        'transitions' => [
+                            ['name' => 'go', 'from' => ['a'], 'to' => ['b']],
+                        ],
+                    ],
+                ],
+            ],
+        ]]);
+    }
+
+    public function testWorkflowEventsToDispatchAcceptsBlockListOnlyList()
+    {
+        $processor = new Processor();
+        $configuration = new Configuration(true);
+
+        $config = $processor->processConfiguration($configuration, [[
+            'http_method_override' => false,
+            'handle_all_throwables' => true,
+            'php_errors' => ['log' => true],
+            'workflows' => [
+                'workflows' => [
+                    'block_list' => [
+                        'supports' => [self::class],
+                        'places' => ['a', 'b'],
+                        'initial_marking' => 'a',
+                        'events_to_dispatch' => ['!workflow.announce'],
+                        'transitions' => [
+                            ['name' => 'go', 'from' => ['a'], 'to' => ['b']],
+                        ],
+                    ],
+                ],
+            ],
+        ]]);
+
+        $this->assertSame(['!workflow.announce'], $config['workflows']['workflows']['block_list']['events_to_dispatch']);
+    }
+
+    public function testWorkflowEventsToDispatchRejectsBlockListedGuardEvent()
+    {
+        $processor = new Processor();
+        $configuration = new Configuration(true);
+
+        $this->expectException(InvalidConfigurationException::class);
+        $this->expectExceptionMessage('The "workflow.guard" event cannot be disabled in "events_to_dispatch": it is always dispatched.');
+
+        $processor->processConfiguration($configuration, [[
+            'http_method_override' => false,
+            'handle_all_throwables' => true,
+            'php_errors' => ['log' => true],
+            'workflows' => [
+                'workflows' => [
+                    'guard_block' => [
+                        'supports' => [self::class],
+                        'places' => ['a', 'b'],
+                        'initial_marking' => 'a',
+                        'events_to_dispatch' => ['!workflow.guard'],
+                        'transitions' => [
+                            ['name' => 'go', 'from' => ['a'], 'to' => ['b']],
+                        ],
+                    ],
+                ],
+            ],
+        ]]);
+    }
+
     public function testFormCsrfProtectionFieldAttrDoNotNormalizeKeys()
     {
         $processor = new Processor();