+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/workflow_with_disabled_events.php
@@ -1,0 +1,43 @@
+<?php
+
+use Symfony\Bundle\FrameworkBundle\Tests\DependencyInjection\FrameworkExtensionTestCase;
+
+$container->loadFromExtension('framework', [
+    'workflows' => [
+        'my_workflow' => [
+            'type' => 'state_machine',
+            'marking_store' => [
+                'property' => 'state',
+            ],
+            'supports' => [
+                FrameworkExtensionTestCase::class,
+            ],
+            'events_to_dispatch' => [
+                '!workflow.announce',
+            ],
+            'places' => [
+                'one',
+                'two',
+                'three',
+            ],
+            'transitions' => [
+                'count_to_two' => [
+                    'from' => [
+                        'one',
+                    ],
+                    'to' => [
+                        'two',
+                    ],
+                ],
+                'count_to_three' => [
+                    'from' => [
+                        'two',
+                    ],
+                    'to' => [
+                        'three',
+                    ],
+                ],
+            ],
+        ],
+    ],
+]);